hyperion::api::flat::message::hyperion_reply_generated::hyperionnet

Function size_prefixed_root_as_reply

pub fn size_prefixed_root_as_reply(
    buf: &[u8],
) -> Result<Reply<'_>, InvalidFlatbuffer>

Verifies that a buffer of bytes contains a size prefixed Reply and returns it. Note that verification is still experimental and may not catch every error, or be maximally performant. For the previous, unchecked, behavior use size_prefixed_root_as_reply_unchecked.