Type Alias hyperion::db::DbError

pub type DbError = Error;

Aliased Type

enum DbError {

    Configuration(Box<dyn Error + Sync + Send, Global>),
    Database(Box<dyn DatabaseError, Global>),
    Io(Error),
    Tls(Box<dyn Error + Sync + Send, Global>),
    Protocol(String),
    RowNotFound,
    TypeNotFound {
        type_name: String,
    },
    ColumnIndexOutOfBounds {
        index: usize,
        len: usize,
    },
    ColumnNotFound(String),
    ColumnDecode {
        index: String,
        source: Box<dyn Error + Sync + Send, Global>,
    },
    Decode(Box<dyn Error + Sync + Send, Global>),
    PoolTimedOut,
    PoolClosed,
    WorkerCrashed,
    Migrate(Box<MigrateError, Global>),

}

Variants

Configuration(Box<dyn Error + Sync + Send, Global>)

Error occurred while parsing a connection string.

Database(Box<dyn DatabaseError, Global>)

Error returned from the database.

Io(Error)

Error communicating with the database backend.

Tls(Box<dyn Error + Sync + Send, Global>)

Error occurred while attempting to establish a TLS connection.

Protocol(String)

Unexpected or invalid data encountered while communicating with the database.

This should indicate there is a programming error in a SQLx driver or there is something corrupted with the connection to the database itself.

RowNotFound

No rows returned by a query that expected to return at least one row.

TypeNotFound

Fields

type_name: String

Type in query doesn’t exist. Likely due to typo or missing user type.

ColumnIndexOutOfBounds

Fields

index: usize

len: usize

Column index was out of bounds.

ColumnNotFound(String)

No column found for the given name.

ColumnDecode

Fields

index: String

source: Box<dyn Error + Sync + Send, Global>

Error occurred while decoding a value from a specific column.

Decode(Box<dyn Error + Sync + Send, Global>)

Error occurred while decoding a value.

PoolTimedOut

A Pool::acquire timed out due to connections not becoming available or because another task encountered too many errors while trying to open a new connection.

PoolClosed

Pool::close was called while we were waiting in Pool::acquire.

WorkerCrashed

A background worker has crashed.

Migrate(Box<MigrateError, Global>)