Module hyperion::effects::runtime

Modules

context

Enums

RuntimeMethodError

Traits

RuntimeMethods

Functions

PyInit_hyperion⚠
This autogenerated function is called by the python interpreter when importing the module.

__pyo3_get_function_abort

__pyo3_get_function_set_color

__pyo3_get_function_set_image

__pyo3_raw_abort⚠

__pyo3_raw_set_color⚠

__pyo3_raw_set_image⚠

abort
Check if the effect should abort execution

do_run

hyperion

hyperion_init

run

set_color
Set a new color for the leds

set_image
Set a new image to process and determine new led colors