Trait hyperion::image::Image

pub trait Image: Sized {
    fn width(&self) -> u16;

    fn height(&self) -> u16;

    fn color_at(&self, x: u16, y: u16) -> Option<Color>;

    unsafe fn color_at_unchecked(&self, x: u16, y: u16) -> Color;

    fn to_raw_image(&self) -> RawImage;
}

Required methods

fn width(&self) -> u16

Get the width of the image, in pixels

fn height(&self) -> u16

Get the height of the image, in pixels

fn color_at(&self, x: u16, y: u16) -> Option<Color>

Get the color at the given coordinates

unsafe fn color_at_unchecked(&self, x: u16, y: u16) -> Color

Get the color at the given coordinates skipping bound checks

fn to_raw_image(&self) -> RawImage

Convert this image trait object to a raw image

Loading content...