Module hyperion::servers::flat::message::hyperion_reply_generated::hyperionnet

Structs

Reply

ReplyArgs

ReplyBuilder

Enums

ReplyOffset

Functions

finish_reply_buffer

finish_size_prefixed_reply_buffer

get_root_as_reply	Deprecated

get_size_prefixed_root_as_reply	Deprecated

root_as_reply
Verifies that a buffer of bytes contains a Reply and returns it. Note that verification is still experimental and may not catch every error, or be maximally performant. For the previous, unchecked, behavior use root_as_reply_unchecked.

root_as_reply_unchecked⚠
Assumes, without verification, that a buffer of bytes contains a Reply and returns it.

root_as_reply_with_opts
Verifies, with the given options, that a buffer of bytes contains a Reply and returns it. Note that verification is still experimental and may not catch every error, or be maximally performant. For the previous, unchecked, behavior use root_as_reply_unchecked.

size_prefixed_root_as_reply
Verifies that a buffer of bytes contains a size prefixed Reply and returns it. Note that verification is still experimental and may not catch every error, or be maximally performant. For the previous, unchecked, behavior use size_prefixed_root_as_reply_unchecked.

size_prefixed_root_as_reply_unchecked⚠
Assumes, without verification, that a buffer of bytes contains a size prefixed Reply and returns it.

size_prefixed_root_as_reply_with_opts
Verifies, with the given verifier options, that a buffer of bytes contains a size prefixed Reply and returns it. Note that verification is still experimental and may not catch every error, or be maximally performant. For the previous, unchecked, behavior use root_as_reply_unchecked.