Module hyperion::api::flat::message::hyperion_request_generated::hyperionnet

Structs

Clear

ClearArgs

ClearBuilder

Color

ColorArgs

ColorBuilder

Command

CommandUnionTableOffset

Image

ImageArgs

ImageBuilder

ImageType

ImageTypeUnionTableOffset

RawImage

RawImageArgs

RawImageBuilder

Register

RegisterArgs

RegisterBuilder

Request

RequestArgs

RequestBuilder

Enums

ClearOffset

ColorOffset

ImageOffset

RawImageOffset

RegisterOffset

RequestOffset

Constants

ENUM_MAX_COMMAND	Deprecated

ENUM_MAX_IMAGE_TYPE	Deprecated

ENUM_MIN_COMMAND	Deprecated

ENUM_MIN_IMAGE_TYPE	Deprecated

ENUM_VALUES_COMMAND	Deprecated

ENUM_VALUES_IMAGE_TYPE	Deprecated

Functions

finish_request_buffer

finish_size_prefixed_request_buffer

get_root_as_request	Deprecated

get_size_prefixed_root_as_request	Deprecated

root_as_request
Verifies that a buffer of bytes contains a Request and returns it. Note that verification is still experimental and may not catch every error, or be maximally performant. For the previous, unchecked, behavior use root_as_request_unchecked.

root_as_request_unchecked⚠
Assumes, without verification, that a buffer of bytes contains a Request and returns it.

root_as_request_with_opts
Verifies, with the given options, that a buffer of bytes contains a Request and returns it. Note that verification is still experimental and may not catch every error, or be maximally performant. For the previous, unchecked, behavior use root_as_request_unchecked.

size_prefixed_root_as_request
Verifies that a buffer of bytes contains a size prefixed Request and returns it. Note that verification is still experimental and may not catch every error, or be maximally performant. For the previous, unchecked, behavior use size_prefixed_root_as_request_unchecked.

size_prefixed_root_as_request_unchecked⚠
Assumes, without verification, that a buffer of bytes contains a size prefixed Request and returns it.

size_prefixed_root_as_request_with_opts
Verifies, with the given verifier options, that a buffer of bytes contains a size prefixed Request and returns it. Note that verification is still experimental and may not catch every error, or be maximally performant. For the previous, unchecked, behavior use root_as_request_unchecked.