hyperion::db

Type Alias DbError

pub type DbError = Error;

Aliased Type

pub enum DbError {

    Configuration(Box<dyn Error + Send + Sync>),
    InvalidArgument(String),
    Database(Box<dyn DatabaseError>),
    Io(Error),
    Tls(Box<dyn Error + Send + Sync>),
    Protocol(String),
    RowNotFound,
    TypeNotFound {
        type_name: String,
    },
    ColumnIndexOutOfBounds {
        index: usize,
        len: usize,
    },
    ColumnNotFound(String),
    ColumnDecode {
        index: String,
        source: Box<dyn Error + Send + Sync>,
    },
    Encode(Box<dyn Error + Send + Sync>),
    Decode(Box<dyn Error + Send + Sync>),
    AnyDriverError(Box<dyn Error + Send + Sync>),
    PoolTimedOut,
    PoolClosed,
    WorkerCrashed,
    Migrate(Box<MigrateError>),
    InvalidSavePointStatement,
    BeginFailed,

}

Variants

Configuration(Box<dyn Error + Send + Sync>)

Error occurred while parsing a connection string.

InvalidArgument(String)

One or more of the arguments to the called function was invalid.

The string contains more information.

Database(Box<dyn DatabaseError>)

Error returned from the database.

Io(Error)

Error communicating with the database backend.

Tls(Box<dyn Error + Send + Sync>)

Error occurred while attempting to establish a TLS connection.

Protocol(String)

Unexpected or invalid data encountered while communicating with the database.

This should indicate there is a programming error in a SQLx driver or there is something corrupted with the connection to the database itself.

RowNotFound

No rows returned by a query that expected to return at least one row.

TypeNotFound

Type in query doesn’t exist. Likely due to typo or missing user type.

Fields

type_name: String

ColumnIndexOutOfBounds

Column index was out of bounds.

Fields

index: usize

len: usize

ColumnNotFound(String)

No column found for the given name.

ColumnDecode

Error occurred while decoding a value from a specific column.

Fields

index: String

source: Box<dyn Error + Send + Sync>

Encode(Box<dyn Error + Send + Sync>)

Error occured while encoding a value.

Decode(Box<dyn Error + Send + Sync>)

Error occurred while decoding a value.

AnyDriverError(Box<dyn Error + Send + Sync>)

Error occurred within the Any driver mapping to/from the native driver.

PoolTimedOut

A Pool::acquire timed out due to connections not becoming available or because another task encountered too many errors while trying to open a new connection.

PoolClosed

Pool::close was called while we were waiting in Pool::acquire.

WorkerCrashed

A background worker has crashed.

Migrate(Box<MigrateError>)

InvalidSavePointStatement

BeginFailed